feat(vm): remove Globals
I forgot what the Globals struct is for, so I am removing it. It doesn't
seem to be used anywhere in the VM. If it turns out to be an important
feature, I can always revert back. The implementation is very easy to
understand. I honestly don't know what I was thinking while writing that
struct.