chore(vm): rename object to types
I feel like this name represent it better. It is filled with structs
that define the *types* and not the object that is stored in memory.
However, I still have no idea how to allocate those types in memory. The
basic heap implementation is already there, but not the types.