chore(devel): remove belalang_devel crate
I thought it would simplify development, but instead it complicates it.
I intended the belalang_devel to be a crate that provides everything
developers need to implement a third party extension. Now that I think
about it, it is just better to use the features feature of crates and
Cargo.